+++ src/UserInformation.h
@@ -7,6 +7,7 @@ struct UserInformation{
     Q_GADGET
     public:
     QString userId;
+    QString cmUserId;
     QString accessToken;
     QString deviceId;
     QString homeServer;
@@ -14,6 +15,7 @@ struct UserInformation{
     QString avatarUrl;
 
     Q_PROPERTY(QString userId MEMBER userId)
+    Q_PROPERTY(QString cmUserId MEMBER cmUserId)
     Q_PROPERTY(QString accessToken MEMBER accessToken)
     Q_PROPERTY(QString deviceId MEMBER deviceId)
     Q_PROPERTY(QString homeServer MEMBER homeServer)